Export `PageLayout` resizable functionality
