+++ .changeset/pagelayout-export-resize-primitives.md
@@ -1,0 +1,5 @@
+---
+'@primer/react': minor
+---
+
+PageLayout: Export the resize primitives `usePaneWidth` and `DragHandle` (with `defaultPaneWidth` and related types) so pane-resize behavior can be reused outside of `PageLayout`.
